requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2        
python-dotenv   
dj-database-url  
django-ckeditor
whitenoise
cloudinary
django-cloudinary-storage
django-extensions
